feat(xplat): add checks for darwin for unique features

+++ src/scripts/flush-dns.sh
@@ -1,4 +1,8 @@
 function flush-dns() {
+	if [ "$(uname)" != "Darwin" ]; then
+		return
+	fi
+	
 	dscacheutil -flushcache
 	sudo killall -HUP mDNSResponder
 }
