+++ src/scripts/hide-all.sh
@@ -1,5 +1,8 @@
 function hide-all() {
+	if [ "$(uname)" != "Darwin" ]; then
+		return
+	fi
+	
 	defaults write com.apple.finder AppleShowAllFiles FALSE
-
 	killall Finder
 }
