+++ src/scripts/show-all.sh
@@ -1,4 +1,8 @@
 function show-all() {
+	if [ "$(uname)" != "Darwin" ]; then
+		return
+	fi
+	
 	defaults write com.apple.finder AppleShowAllFiles TRUE
 	killall Finder
 }
